src/requirements.txt
flask
mangum
asgiref
discord-interactions
python-dotenv